Fix implementation-defined integer conversion
"When an integer is demoted to a signed integer with smaller size, or an
unsigned integer is converted to its corresponding signed integer, if
the value cannot be represented the result is implementation-defined."

+++ library/x509_crt.c
@@ -2182,7 +2182,7 @@ exit:
 
     if( ret != 0 )
     {
-        *flags = (uint32_t) -1;
+        *flags = -1;
         return( ret );
     }
